requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
distlib==0.3.6
Django==3.2.18
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.2.0
django-filter==23.2
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
drf-spectacular-sidecar==2023.3.1
drf-yasg==1.21.5
Faker==17.6.0
filelock==3.12.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysqlclient==2.2.0
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
platformdirs==3.8.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
toposort==1.10
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.23.1
whitenoise==6.4.0
